Remove map limits
- Remove map limits functionality
- Refactor: MapTiles + PersonFilterEnum => utils.py
